+++ src/winmain.c
@@ -31,6 +31,7 @@ static int extra_width, extra_height;
 static bool fullscr_on_max;
 static bool resizing;
 static bool title_settable = true;
+static string border_style = 0;
 
 static HBITMAP caretbm;
 
@@ -385,7 +386,14 @@ clear_fullscreen(void)
 
  /* Reinstate the window furniture. */
   LONG style = GetWindowLong(wnd, GWL_STYLE);
-  style |= WS_CAPTION | WS_BORDER | WS_THICKFRAME;
+  if (border_style) {
+    if (strcmp (border_style, "void") != 0) {
+      style |= WS_THICKFRAME;
+    }
+  }
+  else {
+    style |= WS_CAPTION | WS_BORDER | WS_THICKFRAME;
+  }
   SetWindowLong(wnd, GWL_STYLE, style);
   SetWindowPos(wnd, null, 0, 0, 0, 0,
                SWP_NOMOVE | SWP_NOSIZE | SWP_NOZORDER | SWP_FRAMECHANGED);
@@ -754,7 +762,7 @@ static const char help[] =
   "  -V, --version         Print version information and exit\n"
 ;
 
-static const char short_opts[] = "+:c:eh:i:l:o:p:s:t:T:uw:HV";
+static const char short_opts[] = "+:c:eh:i:l:o:p:s:t:T:B:uw:HV";
 
 static const struct option
 opts[] = {
@@ -769,6 +777,7 @@ opts[] = {
   {"size",     required_argument, 0, 's'},
   {"title",    required_argument, 0, 't'},
   {"Title",    required_argument, 0, 'T'},
+  {"Border",   required_argument, 0, 'B'},
   {"window",   required_argument, 0, 'w'},
   {"class",    required_argument, 0, 'C'},
   {"help",     no_argument,       0, 'H'},
@@ -871,6 +880,8 @@ main(int argc, char *argv[])
       when 'T':
         set_arg_option("Title", optarg);
         title_settable = false;
+      when 'B':
+        border_style = strdup (optarg);
       when 'u': cfg.utmp = true;
       when 'w': set_arg_option("Window", optarg);
       when 'C': set_arg_option("Class", optarg);
@@ -1066,6 +1077,19 @@ main(int argc, char *argv[])
                         WS_OVERLAPPEDWINDOW | (cfg.scrollbar ? WS_VSCROLL : 0),
                         cfg.x, cfg.y, width, height,
                         null, null, inst, null);
+
+  if (border_style) {
+    LONG style = GetWindowLong(wnd, GWL_STYLE);
+    if (strcmp (border_style, "void") == 0) {
+      style &= ~(WS_CAPTION | WS_BORDER | WS_THICKFRAME);
+    }
+    else {
+      style &= ~(WS_CAPTION | WS_BORDER);
+    }
+    SetWindowLong(wnd, GWL_STYLE, style);
+    SetWindowPos(wnd, null, 0, 0, 0, 0,
+                 SWP_NOMOVE | SWP_NOSIZE | SWP_NOZORDER | SWP_FRAMECHANGED);
+  }
 
   // The input method context.
   imc = ImmGetContext(wnd);